[Finishes #163832052] restructuring project folders and files for efficiency
