+++ run.py
@@ -1,7 +1,7 @@
 import os
 
 from app import create_app
-config_name = os.getenv('APP_SETTINGS')  # config_name='development'
+config_name = os.getenv('APP_SETTINGS')
 app = create_app(config_name)
 
 if __name__ == '__main__':
